// templates/gorgias/types.ts
export interface GorgiasUserResponse {
    id: number;
    active: boolean;
    name: string;
    email: string;
    bio?: string | null;
    country?: string | null;
    external_id?: string;
    language?: string | null;
    firstname: string;
    lastname: string;
    meta?: Record<string, any>;
    role: {
        name: 'admin' | 'agent' | 'basic-agent' | 'lite-agent' | 'observer-agent' | 'bot';
    };
    created_datetime: string;
    updated_datetime: string | null;
    deactivated_datetime: string | null;
    timezone: string | null;
}
export interface GorgiasCreateUserReq {
    name: string;
    email: string;
    role: {
        name: 'admin' | 'agent' | 'basic-agent' | 'lite-agent' | 'observer-agent' | 'bot';
    };
}